src/drivesense/inference/serve.py
"""Production serving wrapper for DriveSense-VLM using vLLM.

Implements Phase 3d: wraps the vLLM AsyncLLMEngine with the Qwen3-VL processor
to expose a clean predict() / predict_batch() API. Parses structured JSON from
model output and validates against the DriveSense output schema.

The TensorRT ViT engine (Phase 3c) is loaded separately and pre-processes
image patches before passing visual tokens to vLLM.

Usage (HPC only):
    python -m drivesense.inference.serve --config configs/inference.yaml

Implementation target: Phase 3d
"""

from __future__ import annotations

from pathlib import Path
from typing import TYPE_CHECKING

if TYPE_CHECKING:
    from PIL import Image

# vLLM — HPC only
try:
    from vllm import AsyncLLMEngine, SamplingParams  # type: ignore[import]
except ImportError:
    AsyncLLMEngine = None  # type: ignore[assignment]
    SamplingParams = None  # type: ignore[assignment]


class DriveSenseServer:
    """Production serving wrapper for DriveSense-VLM.

    Loads the quantized model into vLLM, maintains a processor for image
    tokenisation, and exposes synchronous predict() and predict_batch() APIs
    that return validated structured JSON hazard detections.

    Args:
        config_path: Path to configs/inference.yaml.
    """

    def __init__(self, config_path: str | Path) -> None:
        raise NotImplementedError("Phase 3d: load vLLM engine and Qwen3-VL processor from config")

    def predict(self, image: Image.Image) -> dict:
        """Run hazard detection on a single dashcam frame.

        Args:
            image: PIL Image — will be resized to max_image_size from config.

        Returns:
            Structured hazard detection dict:
            ``{"bbox_2d": [x1, y1, x2, y2], "hazard_class": str,
               "severity": int, "reasoning": str, "action": str}``.

        Raises:
            ValueError: If model output cannot be parsed as valid JSON.
        """
        raise NotImplementedError("Phase 3d: preprocess image, call vLLM, parse JSON response")

    def predict_batch(self, images: list[Image.Image]) -> list[dict]:
        """Run hazard detection on a batch of dashcam frames.

        Args:
            images: List of PIL Images.

        Returns:
            List of structured hazard detection dicts in the same order as input.
        """
        raise NotImplementedError("Phase 3d: batch predict using vLLM async engine")

    def _parse_output(self, raw_text: str) -> dict:
        """Parse and validate model output as structured JSON.

        Args:
            raw_text: Raw model generation string.

        Returns:
            Validated hazard detection dict.

        Raises:
            ValueError: If output is not valid JSON or missing required keys.
        """
        raise NotImplementedError("Phase 3d: extract JSON from raw_text and validate schema")
